installd: Preserve symlink of native lib dir unless it's asec
This is for something I'd call lib2ext: https://gist.github.com/1244362
Without this patch, symlinks won't survive a rom update, and then /data
would be blown up (in my position)

+++ cmds/installd/commands.c
@@ -949,7 +949,7 @@ done:
 int linklib(const char* dataDir, const char* asecLibDir)
 {
     char libdir[PKG_PATH_MAX];
-    struct stat s, libStat;
+    struct stat s, libStat, libdirStat;
     int rc = 0;
 
     const size_t libdirLen = strlen(dataDir) + strlen(PKG_LIB_POSTFIX);
@@ -977,6 +977,12 @@ int linklib(const char* dataDir, const char* asecLibDir)
     }
 
     if (lstat(libdir, &libStat) < 0) {
+        LOGE("couldn't lstat lib dir: %s\n", strerror(errno));
+        rc = -1;
+        goto out;
+    }
+
+    if (stat(libdir, &libdirStat) < 0) {
         LOGE("couldn't stat lib dir: %s\n", strerror(errno));
         rc = -1;
         goto out;
@@ -988,6 +994,14 @@ int linklib(const char* dataDir, const char* asecLibDir)
             goto out;
         }
     } else if (S_ISLNK(libStat.st_mode)) {
+        /* TODO: find if there is a better way to check the symlink */
+        if ((libdirStat.st_mode & S_IWUSR) != 0) {
+        /* symlinked to eg. /sd-ext/lib/{...} by user or script */
+            if (delete_dir_contents(libdir, 0, 0) < 0) {
+                rc = -1;
+                goto out;
+            }
+        }
         if (unlink(libdir) < 0) {
             rc = -1;
             goto out;
@@ -1024,7 +1038,7 @@ out:
 int unlinklib(const char* dataDir)
 {
     char libdir[PKG_PATH_MAX];
-    struct stat s, libStat;
+    struct stat s, libStat, libdirStat;
     int rc = 0;
 
     const size_t libdirLen = strlen(dataDir) + strlen(PKG_LIB_POSTFIX);
@@ -1054,6 +1068,12 @@ int unlinklib(const char* dataDir)
     }
 
     if (lstat(libdir, &libStat) < 0) {
+        LOGE("couldn't lstat lib dir: %s\n", strerror(errno));
+        rc = -1;
+        goto out;
+    }
+
+    if (stat(libdir, &libdirStat) < 0) {
         LOGE("couldn't stat lib dir: %s\n", strerror(errno));
         rc = -1;
         goto out;
@@ -1065,9 +1085,21 @@ int unlinklib(const char* dataDir)
             goto out;
         }
     } else if (S_ISLNK(libStat.st_mode)) {
-        if (unlink(libdir) < 0) {
-            rc = -1;
+        /* TODO: find if there is a better way to check the symlink */
+        if ((libdirStat.st_mode & S_IWUSR) != 0) {
+        /* symlinked to eg. /sd-ext/lib/{...} by user or script */
+            if (delete_dir_contents(libdir, 0, 0) < 0) {
+                rc = -1;
+                goto out;
+            }
+            /* skip preparing lib dir */
             goto out;
+        } else {
+        /* symlinked to /mnt/asec/{...} by system */
+            if (unlink(libdir) < 0) {
+                rc = -1;
+                goto out;
+            }
         }
     }
 
@@ -1079,7 +1111,7 @@ int unlinklib(const char* dataDir)
 
     if (chown(libdir, AID_SYSTEM, AID_SYSTEM) < 0) {
         LOGE("cannot chown dir '%s': %s\n", libdir, strerror(errno));
-        unlink(libdir);
+        rmdir(libdir);
         rc = -errno;
         goto out;
     }
